fix: initializer task state tracking and retry flow
