+++ common/architecture_evm.go
@@ -1,0 +1,46 @@
+package common
+
+import (
+	"context"
+)
+
+const (
+	UpstreamTypeEvm UpstreamType = "evm"
+)
+
+type EvmUpstream interface {
+	Upstream
+	EvmGetChainId(ctx context.Context) (string, error)
+	EvmIsBlockFinalized(blockNumber int64) (bool, error)
+	EvmSyncingState() EvmSyncingState
+	EvmStatePoller() EvmStatePoller
+}
+
+type EvmNodeType string
+
+const (
+	EvmNodeTypeFull    EvmNodeType = "full"
+	EvmNodeTypeArchive EvmNodeType = "archive"
+	EvmNodeTypeLight   EvmNodeType = "light"
+)
+
+type EvmSyncingState int
+
+const (
+	EvmSyncingStateUnknown EvmSyncingState = iota
+	EvmSyncingStateSyncing
+	EvmSyncingStateNotSyncing
+)
+
+type EvmStatePoller interface {
+	Bootstrap(ctx context.Context) error
+	SyncingState() EvmSyncingState
+	SetSyncingState(state EvmSyncingState)
+	LatestBlock() int64
+	FinalizedBlock() int64
+	IsBlockFinalized(blockNumber int64) (bool, error)
+	SuggestFinalizedBlock(blockNumber int64)
+	SuggestLatestBlock(blockNumber int64)
+	SetNetworkConfig(cfg *EvmNetworkConfig)
+	IsObjectNull() bool
+}
